requirements.txt
Django==5.0.7
uvicorn==0.30.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.8.2
python-dotenv==1.0.1
aiocache==0.12.2
redis==5.0.7
pytest==8.2.0
pytest-asyncio==0.23.8